requirements.txt
flask
flask-cors
pyjwt
mysqlclient
python-dotenv
flask-mysqldb
